+++ main/src/test/scala/org/clulab/reach/TestPolarity.scala
@@ -1,12 +1,10 @@
 package org.clulab.reach
 
 import org.clulab.reach.TestUtils._
-import org.clulab.reach.mentions.BioEventMention
 import org.scalatest.{FlatSpec, Matchers}
 
 
 class TestPolarity extends FlatSpec with Matchers{
-  val classifierTemp = new org.clulab.polarity.ml.DeepLearningPolarityClassifier()
 
   def activationBehaivor(sentence:String, controller:String, controlled:String, positive:Boolean, ignored:Boolean = false): Unit ={
 
@@ -16,29 +14,6 @@ class TestPolarity extends FlatSpec with Matchers{
       it should description in {
         val mentions = getBioMentions(sentence)
         info(s"Num mentions: ${mentions.size}")
-
-        println("=====================")
-        println(sentence)
-        val bioEventM = mentions filter {
-          case em:BioEventMention => true
-          case _ => false
-        }
-        for (bioM <- bioEventM){
-          if (bioM.isInstanceOf[BioEventMention]){
-            val masked_lemmas = classifierTemp.maskEvent(bioM.sentenceObj.words.clone(), bioM.asInstanceOf[BioEventMention] , "tag")
-            println("\t",masked_lemmas.toSeq)
-
-            if (bioM.arguments.contains("controller")){
-              println("\t\tcontroller: ",bioM.arguments("controller").head.text, "\ttype:", bioM.arguments("controller").head.getClass.getName, bioM.arguments("controller").head.start, bioM.arguments("controller").head.end)
-            }else {println("\t\tNo controller")}
-            if (bioM.arguments.contains("controlled")){
-              println("\t\tcontrolled: ",bioM.arguments("controlled").head.text, "\ttype:", bioM.arguments("controlled").head.getClass.getName, bioM.arguments("controlled").head.start, bioM.arguments("controlled").head.end)
-            }else {println("\t\tNo controlled")}
-          }
-
-        }
-        scala.io.StdIn.readLine()
-
 
         val result =
           if (positive)
