+++ main/src/test/scala/org/clulab/reach/TestUtils.scala
@@ -187,6 +187,29 @@ object TestUtils {
                             controlledArgs: Seq[String],
                             mentions: Seq[Mention]): Boolean = {
 
+    println("=====================")
+    val classifierTemp = new org.clulab.polarity.ml.DeepLearningPolarityClassifier()
+    println(mentions(0).sentenceObj.words)
+    val bioEventM = mentions filter {
+      case em:BioEventMention => true
+      case _ => false
+    }
+    for (bioM <- bioEventM){
+      if (bioM.isInstanceOf[BioEventMention]){
+        val masked_lemmas = classifierTemp.maskEvent(bioM.sentenceObj.words.clone(), bioM.asInstanceOf[BioEventMention] , "tag")
+        println("\t",masked_lemmas.toSeq)
+
+        if (bioM.arguments.contains("controller")){
+          println("\t\tcontroller: ",bioM.arguments("controller").head.text, "\ttype:", bioM.arguments("controller").head.getClass.getName, bioM.arguments("controller").head.start, bioM.arguments("controller").head.end)
+        }else {println("\t\tNo controller")}
+        if (bioM.arguments.contains("controlled")){
+          println("\t\tcontrolled: ",bioM.arguments("controlled").head.text, "\ttype:", bioM.arguments("controlled").head.getClass.getName, bioM.arguments("controlled").head.start, bioM.arguments("controlled").head.end)
+        }else {println("\t\tNo controlled")}
+      }
+    }
+    scala.io.StdIn.readLine()
+
+
     for (m <- mentions) {
       if (!m.isInstanceOf[TextBoundMention]) {
         if (m.labels contains label) {
